fix: Possible fix for manifest download fail
It is unclear why attempting to download the manifest can result in a rare 500 error. 

This attempts to fix the issue by passing the api key the error says is missing. It also refactors the code to remove a duplicate function called getBungieDefinition.

Finally getJSONblob was moved to stop a circular dependency warning.
